fixed some minor problems , added UserServiceClass some change in ScoreServiceClass

+++ src/main/java/com/github/zmilad97/onlineExam/services/CorrectionService.java
@@ -37,7 +37,7 @@ public class CorrectionService {
             scores.setUserId(user.getId());
             scores.setExamId(exam.getId());
             scores.setQuestionId(Long.parseLong(String.valueOf(questions.get(i).getId())));
-            scores.setAnswer(Integer.parseInt(answersMap.get(questions.get(i).getId())));
+            scores.setAnswer(Long.parseLong(answersMap.get(questions.get(i).getId())));
             scoreService.save(scores);
         }
